This adds a class for a collection response holding only IDs
According to [https://jsonapi.org/format/1.1/#document-top-level], the
top-level `data` member may consist of an array of resource IDs (which
are structurally identical to a resource object with all optional
fields omitted). This is useful, for example, as a return value for
identifying some objects (Dispatches) by relationship without pulling
their properties separately.