CI version 0.2.4 (build 24)
https://github.com/vendorflow/vendorflow-jsonapi/pull/9 This adds a class for a collection response holding only IDs

According to [https://jsonapi.org/format/1.1/#document-top-level], the
top-level  member may consist of an array of resource IDs (which
are structurally identical to a resource object with all optional
fields omitted). This is useful, for example, as a return value for
identifying some objects (Dispatches) by relationship without pulling
their properties separately.

[version patch]